app/runtime/requirements.txt
numpy
scikit-learn
torch
sentence-transformers
faiss-gpu
networkx
tqdm
fastapi
uvicorn
pydantic